Comment out replaceAll for selfAssessment to work
textBoxContent = textBoxContent.replaceAll("’", "'"); is used to improve behavior on mobile, where sometimes a different quotation mark is used (which wouldn't be recognized for e.g. I'm happy, and thus we couldn't return feedback to the user about those contractions used).
TODO: Find fix instead of commenting out. https://www.designcise.com/web/tutorial/how-to-fix-replaceall-is-not-a-function-javascript-error might help.